Touchscreen enable/disable, mirroring the touchpad toggle
Same driver-free path as the touchpad: disable the parent I2C-HID node via
SetupAPI/CfgMgr32 — soft CM_Disable first, persistent SetupAPI fallback if
vetoed, and self-heal on boot so it never sticks off.

The touchpad's device-control logic is extracted verbatim into a shared
HidNodeToggle base; TouchpadControl and the new TouchscreenControl differ
only in the HID collection (touchpad U:0005, touchscreen U:0004) and where
they cache the node id — touchpad behaviour (incl. its log strings) is
preserved by construction.

Full parity with the touchpad surface: quick-panel cell (shown only when a
touchscreen is present), OSD flash, remappable key action, Settings toggle,
RU/EN/ZH strings. New OSD icons reuse the touchpad tap-gesture hand + ripple
with a screen bezel so the pair reads as one family.
