+++ README.md
@@ -19,8 +19,8 @@
 </p>
 
 *Панель быстрых настроек (удержание кнопки Mi): пять режимов производительности,
-лимит заряда, «В дорогу» (разовый заряд до 100%), тачпад вкл/выкл, авто-герцовка
-и «режим совы» (не спать).*
+лимит заряда, «В дорогу» (разовый заряд до 100%), тачпад и сенсорный экран вкл/выкл,
+авто-герцовка и «режим совы» (не спать).*
 
 <p align="center">
   <img src="docs/img/monitor.png" width="440" alt="Окно «Монитор»">
@@ -68,6 +68,9 @@
 - 🖱️ **Тачпад вкл/выкл** — действие для любой клавиши + ячейка в панели. Отключение
   штатное (как в Диспетчере устройств, без драйверов) и не переживает перезагрузку —
   залипнуть выключенным тачпад не может.
+- 👆 **Сенсорный экран вкл/выкл** — то же самое для тачскрина ноутбука: действие для
+  клавиши, ячейка в панели, штатное отключение без драйверов и авто-включение после
+  перезагрузки. Ячейка появляется только если сенсорный экран в системе есть.
 - ⚙️ **Окно «Настройки»** в стиле Windows 11 — все опции по вкладкам (Общие / Батарея /
   Экран / Производительность / Клавиши / О программе), тёмная и светлая темы.
 - 🎨 Значок в трее меняется по режиму, монохром под светлую/тёмную панель задач;
@@ -142,7 +145,7 @@ dotnet publish src/XiControl.csproj -c Release -r win-x64 --self-contained -p:Pu
 | Клавиша подсветки клавиатуры | OSD с уровнем (выкл / 50% / 100% / авто) |
 
 Все опции собраны в окне **«Настройки…»** (пункт меню трея): вкладки Общие (язык,
-автозапуск, запоминание яркости, сова, тачпад), Батарея (звук «В дорогу»), Экран
+автозапуск, запоминание яркости, сова, тачпад, сенсорный экран), Батарея (звук «В дорогу»), Экран
 (авто-герцовка и частоты), Производительность (видимость режимов, режим при старте,
 профили питания), Клавиши (переназначение) и О программе. Быстрые переключатели
 (заряд, «В дорогу», сова, герцовка, «Монитор», режим) остаются в меню трея и панели.
@@ -274,12 +277,13 @@ dotnet publish src/XiControl.csproj -c Release -r win-x64 --self-contained -p:Pu
 Каждой клавише — своё действие: **Настройки → Клавиши**. Слоты — одиночный и двойной
 клик Mi-кнопки, клавиши «Настройки» (шестерёнка), AI и «Проекция». На любой слот можно
 навесить: цикл режимов, заряд 80/100, быструю панель, режим совы, «Монитор»,
-«В дорогу», тачпад вкл/выкл, системные «Проекция (Win+P)» / «Параметры Windows» /
-«Copilot (Win+C)», запуск своей программы или «Ничего».
+«В дорогу», тачпад и сенсорный экран вкл/выкл, системные «Проекция (Win+P)» /
+«Параметры Windows» / «Copilot (Win+C)», запуск своей программы или «Ничего».
 
-Тачпад отключается штатным механизмом Windows (как «Отключить устройство» в Диспетчере,
-без драйверов) и **всегда включается сам после перезагрузки** — залипнуть выключенным
-не может. Ячейка тачпада есть и в быстрой панели, рядом с авто-герцовкой.
+Тачпад и сенсорный экран отключаются штатным механизмом Windows (как «Отключить
+устройство» в Диспетчере, без драйверов) и **всегда включаются сами после перезагрузки** —
+залипнуть выключенными не могут. Их ячейки есть и в быстрой панели, рядом с авто-герцовкой
+(ячейка сенсорного экрана — только если тачскрин в системе присутствует).
 
 - Удержание Mi-кнопки всегда открывает быструю панель (не настраивается).
 - Двойной клик Mi = «Ничего» → жест отключён, одиночный клик срабатывает мгновенно
@@ -360,7 +364,7 @@ built and tested primarily on the **Xiaomi Book Pro 14 (2026)** but not limited 
 battery charge limit (~80% / 100%) with automatic re-arm after sleep, a one-off
 "travel charge" to 100%, performance modes (Eco / Quiet / Auto / Turbo / Full speed),
 OSD overlays, auto refresh rate and brightness memory per power source, touchpad
-on/off, a Windows 11-style settings window, remappable Mi button and special keys
+and touchscreen on/off, a Windows 11-style settings window, remappable Mi button and special keys
 (any function on any key, including launching your own program), and fixes for
 otherwise dead keys.
