+++ src/Config/AppConfig.cs
@@ -175,6 +175,22 @@ public sealed class AppConfig
     /// старте включает тачпад само. Ставится/снимается автоматически.</summary>
     public bool TouchpadPersistOff { get; set; }
 
+    /// <summary>
+    /// «Управление сенсорным экраном» как фича: ячейка в панели и действие для клавиш.
+    /// false — ячейка скрыта, действие «touchscreen» не срабатывает (как OwlMode у совы).
+    /// </summary>
+    public bool TouchscreenFeature { get; set; } = true;
+
+    /// <summary>ID узла сенсорного экрана — запоминается автоматически при первом обнаружении.
+    /// Нужен, чтобы включить экран обратно после перезапуска приложения: у выключенного
+    /// экрана HID-коллекции исчезают из системы и найти его иначе нечем. Не редактировать.</summary>
+    public string? TouchscreenDeviceId { get; set; }
+
+    /// <summary>Сенсорный экран отключён персистентным путём (мягкое отключение не сработало —
+    /// фолбэк через SetupAPI, переживает перезагрузку). По этому флагу приложение на
+    /// старте включает экран само. Ставится/снимается автоматически.</summary>
+    public bool TouchscreenPersistOff { get; set; }
+
     // ---- Устаревшие поля клавиш (до v0.8) — читаются только для миграции ----
 
     /// <summary>Устаревшее: см. SettingsKeyAction. "charge"/"settings".</summary>
